Inital Commit
In this VR-based project, the player controls a robot character. The goal is to learn the skills of collecting by doing the squat movement correctly. The system gives a warning when the player leans incorrectly. In order to successfully complete the task, he needs to pick up the baskets from the ground and collect the right amount to feed the beaver. The tasks completed with the right technique and amount are successful.